requirements.txt
# Bitcoin Ticker Dependencies
PyQt6>=6.6.0
websocket-client>=1.6.0
onnxruntime>=1.16.0
numpy>=1.24.0
soundfile>=0.12.0
requests>=2.31.0
